RADX-628 | updated code
