+++ src/Command/DrupalUpdate/DrupalPackageManager.php
@@ -2,12 +2,13 @@
 
 namespace Acquia\Cli\Command\DrupalUpdate;
 
+use Acquia\Cli\Exception\AcquiaCliException;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
-class DrupalPackagesInfo {
-
+class DrupalPackageManager {
   /**
    * @var OutputInterface
    */
@@ -22,6 +23,21 @@ class DrupalPackagesInfo {
    * @var bool
    */
   private bool $isCoreUpdated;
+
+  /**
+   * @var SymfonyStyle
+   */
+  private SymfonyStyle $io;
+
+  /**
+   * @var array
+   */
+  public array $infoPackageFilesPath=[];
+
+  /**
+   * @var FileSystemUtility
+   */
+  private FileSystemUtility $fileSystemUtility;
 
   /**
    * @var DrupalOrgClient
@@ -63,10 +79,19 @@ class DrupalPackagesInfo {
     $this->output = $output;
   }
 
+  /**
+   * @param FileSystemUtility $fileSystemUtility
+   */
+  public function setFileSystemUtility(FileSystemUtility $fileSystemUtility): void {
+    $this->fileSystemUtility = $fileSystemUtility;
+  }
+
   public function __construct(InputInterface $input, OutputInterface $output) {
     $this->setIsCoreUpdated(FALSE);
     $this->setDrupalOrgClient(new DrupalOrgClient($input, $output));
     $this->setOutput($output);
+    $this->io = new SymfonyStyle($input, $output);
+    $this->setFileSystemUtility(new FileSystemUtility($input, $output));
   }
 
   /**
@@ -88,6 +113,7 @@ class DrupalPackagesInfo {
   /**
    * @param array $info_packages_file_path
    * @param string $drupal_project_cwd_path
+   * @throws AcquiaCliException
    */
   public function getPackageDetailInfo(array $info_packages_file_path, string $drupal_project_cwd_path): void {
     foreach ($info_packages_file_path as $package_name => $package_path) {
@@ -111,13 +137,13 @@ class DrupalPackagesInfo {
   public function getFileInfo(string $filepath, string $drupal_project_cwd_path): void {
     $drupal_client = $this->getDrupalOrgClient();
     $package_info_key = [
-      'name',
-      'description',
-      'package',
-      'version',
-      'core',
-      'project',
-    ];
+          'name',
+          'description',
+          'package',
+          'version',
+          'core',
+          'project',
+      ];
     $info_extension_file = @parse_ini_file($filepath, FALSE, INI_SCANNER_RAW);
     if (is_bool($info_extension_file) && $info_extension_file == FALSE) {
       return;
@@ -184,13 +210,105 @@ class DrupalPackagesInfo {
       $git_commit_message_detail[] = $git_commit_message;
     }
     $table->setHeaders([
-       'Package Name',
-       'Package Type',
-       'Current Version',
-       'Latest Version',
-       'Update Type',
-     ])->setRows($git_commit_message_detail);
+          'Package Name',
+          'Package Type',
+          'Current Version',
+          'Latest Version',
+          'Update Type',
+      ])->setRows($git_commit_message_detail);
     $table->render();
+  }
+
+  /**
+   * Get Packages detail information.
+   * Package name, package type, package current version etc.
+   * @param string $drupal_project_cwd_path
+   * @return array
+   * @throws AcquiaCliException
+   */
+  public function getPackagesMetaData(string $drupal_project_cwd_path): array {
+    $this->io->note('Checking available updates.');
+    $this->infoPackageFilesPath = $this->fileSystemUtility->getInfoFilesList($drupal_project_cwd_path);
+    if (count($this->infoPackageFilesPath) == 0) {
+      throw new AcquiaCliException("Not valid Drupal 7 project.");
+    }
+    $this->io->note('Preparing packages.');
+    $this->getPackageDetailInfo($this->infoPackageFilesPath, $drupal_project_cwd_path);
+    return $this->availablePackageUpdatesList($drupal_project_cwd_path);
+  }
+
+  /**
+   * @param string $drupal_project_cwd_path
+   * @return array
+   */
+  public function availablePackageUpdatesList(string $drupal_project_cwd_path): array {
+    $version_detail = $this->getPackageData();
+    $package_info_files = $this->infoPackageFilesPath;
+    $drupal_docroot_path = $drupal_project_cwd_path . '/docroot';
+    $git_commit_message_detail = [];
+    $git_commit_message_detail[] = [
+          'Package Name',
+          'Package Type',
+          'Current Version',
+          'Latest Version',
+          'Update Type',
+          'Download Link',
+          'File Path',
+      ];
+    foreach ($version_detail as $package => $versions) {
+      if (!isset($versions['available_versions'])) {
+        continue;
+      }
+      $git_commit_message['package'] = $package;
+      $git_commit_message['package_type'] = $versions['package_type'];
+      $git_commit_message['current_version'] = $versions['current_version'] ?? '';
+      $git_commit_message['latest_version'] = $versions['available_versions']['version'] ?? '';
+      $git_commit_message['update_notes'] = isset($versions['available_versions']['terms']) ? $this->getUpdateType($versions['available_versions']['terms']['term']) : '';
+      $git_commit_message['download_link'] = $versions['available_versions']['download_link'] ?? '';
+      if (isset($package_info_files[$package . '.info']) && str_contains($package_info_files[$package . '.info'], ",")) {
+        $package_info_files[$package . '.info'] = explode(',', $package_info_files[$package . '.info']);
+      }
+
+      if (isset($package_info_files[$package . '.info']) && is_array($package_info_files[$package . '.info'])) {
+        $file_paths=[];
+        foreach ($package_info_files[$package . '.info'] as $p => $path_location) {
+          $file_path_temp = isset($path_location) ? (str_replace($package . '/' . $package . '.info', '', $path_location)) : '';
+          if (($file_path_temp == '') && ($versions['package_type'] == 'module')) {
+            $file_paths[] = $drupal_docroot_path . "/sites/all/modules";
+          }
+          else {
+            $file_paths[] = ($file_path_temp != '') ? realpath($file_path_temp) : $drupal_docroot_path;
+          }
+        }
+        $git_commit_message['file_path'] = $file_paths;
+      }
+      else {
+        $file_path = isset($package_info_files[$package . '.info']) ? (str_replace($package . '/' . $package . '.info', '', $package_info_files[$package . '.info'])) : '';
+        $git_commit_message['file_path'] = ($file_path != '') ? realpath($file_path) : $drupal_docroot_path;
+        if (($file_path == '') && ($versions['package_type'] == 'module')) {
+          $git_commit_message['file_path'] = ($file_path != '') ? realpath($file_path) : $drupal_docroot_path . "/sites/all/modules";
+        }
+      }
+
+      $git_commit_message_detail[] = $git_commit_message;
+    }
+    return $git_commit_message_detail;
+  }
+
+  /**
+   * Get package release type from available release response.
+   * Ex. bug fix, new feature, security update etc.
+   * @param array $package_release_detail
+   * @return string
+   */
+  protected function getUpdateType(array $package_release_detail): string {
+    if (isset($package_release_detail[0]['value'])) {
+      return $package_release_detail[0]['value'];
+    }
+    elseif (isset($package_release_detail['value'])) {
+      return $package_release_detail['value'];
+    }
+    return "";
   }
 
 }
